fix: code review issues
