+++ cmd/devstream/init.go
@@ -1,20 +1,12 @@
 package main
 
 import (
-	"fmt"
-	"os"
-	"os/exec"
-	"path/filepath"
-
 	"github.com/spf13/cobra"
 
 	"github.com/merico-dev/stream/internal/pkg/configloader"
 	"github.com/merico-dev/stream/internal/pkg/pluginmanager"
 	"github.com/merico-dev/stream/pkg/util/log"
-	md5helper "github.com/merico-dev/stream/pkg/util/md5"
 )
-
-const DTMRemoteMD5 = ".remote"
 
 var initCMD = &cobra.Command{
 	Use:   "init",
@@ -30,7 +22,7 @@ func initCMDFunc(cmd *cobra.Command, args []string) {
 	}
 
 	log.Info("Initialize started.")
-	if err := compareDtmMD5(); err != nil {
+	if err := pluginmanager.CompareDtmMD5(); err != nil {
 		log.Error(err)
 		return
 	}
@@ -42,36 +34,4 @@ func initCMDFunc(cmd *cobra.Command, args []string) {
 	}
 
 	log.Success("Initialize finished.")
-}
-
-// compareDtmMD5 compare dtm itself and remote md5, if does not match, cannot  download plugins from remote
-func compareDtmMD5() error {
-	// get remote dtm .md5
-	dtmMD5FileName := configloader.GetDtmMD5FileName()
-	if err := pluginmanager.DownloadDtmMD5(DTMRemoteMD5, dtmMD5FileName); err != nil {
-		return err
-	}
-
-	selfDtmName, err := getDtmSelf()
-	if err != nil {
-		return err
-	}
-
-	isMD5Match, err := md5helper.ValidateFileMatchMD5(selfDtmName, filepath.Join(DTMRemoteMD5, dtmMD5FileName))
-	if err != nil {
-		return err
-	}
-	if !isMD5Match {
-		return fmt.Errorf("DTM %s doesn't match with remote .md5, cannot download Plugins from remote release version", selfDtmName)
-	}
-	return nil
-}
-
-// getDtmSelf get executing dtm itself
-func getDtmSelf() (string, error) {
-	path, err := exec.LookPath(os.Args[0])
-	if err != nil {
-		return "", err
-	}
-	return path, nil
 }
