+++ internal/pkg/md5/md5.go
@@ -1,4 +1,4 @@
-package version
+package md5
 
 import (
 	"io/ioutil"
@@ -8,8 +8,8 @@ import (
 
 const md5_length = 32
 
-// ValidateFileMatchMD5 check current PlugIn MD5 matches with .md5 file
-func ValidateFileMatchMD5(fileName, md5FileName string) (bool, error) {
+// FileMatchesMD5 checks current PlugIn MD5 matches with .md5 file
+func FileMatchesMD5(fileName, md5FileName string) (bool, error) {
 	currentPlugInMD5, err := CalcFileMD5(fileName)
 	if err != nil {
 		return false, err
@@ -19,7 +19,7 @@ func ValidateFileMatchMD5(fileName, md5FileName string) (bool, error) {
 	if err != nil {
 		return false, err
 	}
-	// intercept string, md5 code is from 0 to 31
+	// intercept string, md5 code length is 32
 	md5Content := strings.ShortenString(string(md5ContentBytes), md5_length)
 
 	return currentPlugInMD5 == md5Content, nil
